server/requirements.txt
asgiref==3.8.1
axios==0.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
diff-match-patch==20241021
Django==5.0.7
django-cors-headers==4.4.0
django-filter==24.3
django-import-export==4.3.5
djangorestframework==3.15.2
idna==3.7
import-export==0.3.1
lxml==5.2.2
markdown-it-py==3.0.0
mdurl==0.1.2
Pygments==2.18.0
requests==2.32.3
rich==13.7.1
sqlparse==0.5.1
tablib==3.8.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
